MVVM Setup
