+++ DubDubGrub/Views/Screens/LocationMapView/LocationMapView.swift
@@ -1,0 +1,44 @@
+//
+//  LocationMapView.swift
+//  DubDubGrub
+//
+//  Created by Carlos Reyes on 6/17/25.
+//
+
+import SwiftUI
+import MapKit
+
+struct LocationMapView: View {
+    
+    @StateObject private var viewModel = LocationMapViewModel()
+
+    var body: some View {
+        ZStack {
+            Map(position: $viewModel.cameraPosition).ignoresSafeArea()
+            
+            VStack {
+                LogoView().shadow(radius: 10)
+                Spacer()
+            }
+        }
+        .alert(item: $viewModel.alertItem, content: { alertItem in
+            Alert(title: alertItem.title, message: alertItem.message, dismissButton: alertItem.dismissButton)
+        })
+        .onAppear {
+            viewModel.getLocations()
+        }
+    }
+}
+
+#Preview {
+    LocationMapView()
+}
+
+struct LogoView: View {
+    var body: some View {
+        Image(.ddgMapLogo)
+            .resizable()
+            .scaledToFit()
+            .frame(height: 70)
+    }
+}
